Add files via upload
Added the code that shows the altitudinal evolution of the sampled surfaces according to the models with the fastest and the slowest glacial erosion rates. These two extreme scenarios are shown for each group of samples that share the same base level.
A call to plot_landscape_evolution.m is added at the end of plot_results.m. Uncomment this line if needed!